* avoid the most common case for an exception: the file does not exist
git-svn-id: http://dev.zenoss.com/svnint/trunk/core@21532 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ Products/ZenUtils/Watchdog.py
@@ -129,7 +129,8 @@ class Watcher:
 
     def _runOnce(self):
         try:
-            os.unlink(self.socketPath)
+            if os.path.exists(self.socketPath):
+                os.unlink(self.socketPath)
         except OSError:
             log.exception("Problem removing old socket %s" % self.socketPath)
         cmd = self.cmd + ['--watchdogPath', self.socketPath]
